A `get-controller` mechanism for modules
