+++ lib/client/email/email.js
@@ -10,4 +10,4 @@ const {
 
 module.exports = EMAIL_URL
   ? new FBEmailClient(SERVICE_SECRET, SERVICE_SLUG, EMAIL_URL) // initialise email client
-  : { send: () => {} } // no op methods for email client
+  : { sendMessage: () => {} } // no op methods for email client
