+++ lib/client/sms/sms.js
@@ -10,4 +10,4 @@ const {
 
 module.exports = SMS_URL
   ? new FBSMSClient(SERVICE_SECRET, SERVICE_SLUG, SMS_URL) // initialise sms client
-  : { send: () => {} } // no op methods for sms client
+  : { sendMessage: () => {} } // no op methods for sms client
