+++ lib/middleware/routes-metadata/routes-metadata.js
@@ -54,6 +54,11 @@ const {
   kludgeUpdates
 } = require('~/fb-runner-node/page/page')
 
+const {
+  hasController: hasModuleController,
+  getController: getModuleController
+} = require('~/fb-runner-node/module/savereturn/controller/page/get-controller')
+
 const getPageController = require('~/fb-runner-node/controller/page/get-controller')
 const getComponentController = require('~/fb-runner-node/controller/component/get-controller')
 
@@ -416,7 +421,12 @@ async function pageHandler (req, res) {
     }
 
     if (!EDITMODE) {
-      pageInstance = await executePreFlight(pageInstance, userData, POST)
+      if (hasModuleController(route)) {
+        pageInstance = await getModuleController(route)
+          .preFlight(pageInstance, userData, POST)
+      } else {
+        pageInstance = await executePreFlight(pageInstance, userData, POST)
+      }
 
       pageInstance = skipPage(pageInstance, userData)
 
@@ -433,7 +443,13 @@ async function pageHandler (req, res) {
     }
 
     if (!EDITMODE) {
-      pageInstance = await executeSetContents(pageInstance, userData, POST)
+      if (hasModuleController(route)) {
+        pageInstance = await getModuleController(route)
+          .setContents(pageInstance, userData, POST)
+      } else {
+        pageInstance = await executeSetContents(pageInstance, userData, POST)
+      }
+
       if (hasRedirect(pageInstance)) return handleRedirect(res, pageInstance, userData)
     }
 
@@ -489,7 +505,12 @@ async function pageHandler (req, res) {
         // validate inbound values
         pageInstance = validateInput(pageInstance, userData)
 
-        pageInstance = await executePostValidation(pageInstance, userData)
+        if (hasModuleController(route)) {
+          pageInstance = await getModuleController(route)
+            .postValidation(pageInstance, userData)
+        } else {
+          pageInstance = await executePostValidation(pageInstance, userData)
+        }
       }
 
       // VERBOSE x 3 - and why not before validation?
@@ -509,7 +530,12 @@ async function pageHandler (req, res) {
 
     if (isConfirmationPage(pageInstance)) clearSessionCookie(res)
 
-    pageInstance = await executePreUpdateContents(pageInstance, userData)
+    if (hasModuleController(route)) {
+      pageInstance = await getModuleController(route)
+        .preUpdateContents(pageInstance, userData)
+    } else {
+      pageInstance = await executePreUpdateContents(pageInstance, userData)
+    }
 
     if (hasRedirect(pageInstance)) {
       await userData.saveData()
@@ -572,7 +598,12 @@ async function pageHandler (req, res) {
     // TODO - move to editor
     pageInstance = await setEditorControls(pageInstance, userData)
 
-    pageInstance = await executePreRender(pageInstance, userData)
+    if (hasModuleController(route)) {
+      pageInstance = await getModuleController(route)
+        .preRender(pageInstance, userData)
+    } else {
+      pageInstance = await executePreRender(pageInstance, userData)
+    }
 
     // Pass userdata to context
     const userdata = userData.getUserData()
