--- lib/module/savereturn/controller/client/save-return.js
@@ -1,26 +1,0 @@
-require('@ministryofjustice/module-alias/register-module')(module)
-
-const FBSaveReturnClient = require('./fb-save-return-client')
-
-const {
-  SERVICE_SLUG,
-  SAVE_RETURN_URL,
-  SERVICE_SECRET,
-  ENCODED_PRIVATE_KEY
-} = require('~/fb-runner-node/constants/constants')
-
-// no op methods for save return client
-const noop = () => {}
-let saveReturnClient = {
-  encrypt: noop,
-  decrypt: noop,
-  createSetupEmailToken: noop,
-  validateSetupEmailToken: noop
-}
-
-// initialise save return client
-if (SAVE_RETURN_URL) {
-  saveReturnClient = new FBSaveReturnClient(SERVICE_SECRET, SERVICE_SLUG, SAVE_RETURN_URL, { encodedPrivateKey: ENCODED_PRIVATE_KEY })
-}
-
-module.exports = saveReturnClient
