--- lib/module/savereturn/controller/controller.js
@@ -1,5 +1,0 @@
-const pageInstance = require('./page/instance/page-instance-controllers')
-
-module.exports = {
-  pageInstance
-}
