--- lib/module/savereturn/controller/page/instance/return.setup.email/return.setup.email.controller.js
@@ -1,20 +1,0 @@
-require('@ministryofjustice/module-alias/register-module')(module)
-
-const { client, sendEmail, getConfig } = require('~/fb-runner-node/module/savereturn/controller/savereturn')
-
-const ReturnEmailController = {}
-
-ReturnEmailController.postValidation = async (pageInstance, userData) => {
-  const userId = userData.getUserId()
-  const userToken = userData.getUserToken()
-  const email = userData.getUserDataProperty('email')
-  const duration = getConfig('emailTokenDuration')
-
-  const token = await client.createSetupEmailToken(userId, userToken, email, duration, userData.logger)
-
-  await sendEmail('email.return.setup.email.token', userData, { token })
-
-  return pageInstance
-}
-
-module.exports = ReturnEmailController
