+++ lib/module/savereturn/controller/page/type/return.setup.email.token/return.setup.email.token.controller.js
@@ -1,0 +1,39 @@
+require('@ministryofjustice/module-alias/register-module')(module)
+
+const cloneDeep = require('lodash.clonedeep')
+
+const CommonController = require('~/fb-runner-node/module/savereturn/controller/page/common')
+
+const {
+  client,
+  handleValidationError,
+  resetUser,
+  authenticate,
+  sendEmail
+} = require('~/fb-runner-node/module/savereturn/controller/savereturn')
+
+module.exports = class SetupEmailTokenController extends CommonController {
+  async preUpdateContents (instance, userData) {
+    const pageInstance = cloneDeep(instance)
+    const token = userData.getUserParam('token')
+    const passphrase = userData.getUserParam('passphrase')
+    try {
+      const details = await client.validateSetupEmailToken(token, passphrase, userData.logger)
+
+      await resetUser(userData, details)
+
+      // unless mandatory 2fa
+      const { userId, userToken, email } = details
+
+      await client.createRecord(userId, userToken, email, undefined, userData.logger)
+      await sendEmail('email.return.setup.email.verified', userData)
+
+      authenticate(userData, 'setup-email')
+
+      pageInstance.redirect = 'return.setup.success'
+    } catch (e) {
+      return handleValidationError(e, 'return.setup.email')
+    }
+    return pageInstance
+  }
+}
