+++ lib/module/savereturn/controller/savereturn.js
@@ -2,12 +2,18 @@ require('@ministryofjustice/module-alias/register-module')(module)
 
 const emailClient = require('~/fb-runner-node/client/email/email')
 const smsClient = require('~/fb-runner-node/client/sms/sms')
+const savereturnClient = require('~/fb-runner-node/module/savereturn/client/savereturn/savereturn')
 
-const client = require('./client/save-return')
-const { getInstance, getInstanceProperty } = require('~/fb-runner-node/service-data/service-data')
-const { format } = require('~/fb-runner-node/format/format')
+const {
+  getInstance,
+  getInstanceProperty
+} = require('~/fb-runner-node/service-data/service-data')
 
-const handleValidationError = (e, type) => {
+const {
+  format
+} = require('~/fb-runner-node/format/format')
+
+function handleValidationError (e, type) {
   if (e.code === 401 && e.message) {
     const errorInstance = getInstance(`${type}.${e.message}`)
     if (errorInstance) {
@@ -17,7 +23,7 @@ const handleValidationError = (e, type) => {
   throw e
 }
 
-const resetUser = async (userData, details) => {
+async function resetUser (userData, details) {
   const { userId, userToken, email, mobile } = details
   await userData.resetUserData(userId, userToken)
   userData.setUserDataProperty('email', email)
@@ -26,19 +32,19 @@ const resetUser = async (userData, details) => {
   }
 }
 
-const authenticate = async (userData, action) => {
+async function authenticate (userData, action) {
   userData.setUserDataProperty('authenticated', true)
   const loginTime = Date.now()
   userData.setUserDataProperty('loginTime', loginTime)
   setHistory(userData, action)
 }
 
-const deauthenticate = (userData) => {
+function deauthenticate (userData) {
   userData.unsetUserDataProperty('authenticated')
   setHistory(userData, 'deauthenticate')
 }
 
-const setHistory = (userData, action) => {
+function setHistory (userData, action) {
   const history = userData.getUserDataProperty('history') || []
   history.push({
     phase: action,
@@ -49,7 +55,7 @@ const setHistory = (userData, action) => {
 
 const getConfig = (prop, defaultValue) => getInstanceProperty('module.savereturn.config', prop, defaultValue)
 
-const getMessage = (_id, userData, dataArgs = {}, personalisation = {}) => {
+function getMessage (_id, userData, dataArgs = {}, personalisation = {}) {
   const message = getInstance(_id)
 
   const data = Object.assign({}, userData.getScopedUserData(), dataArgs)
@@ -68,7 +74,7 @@ const getMessage = (_id, userData, dataArgs = {}, personalisation = {}) => {
 }
 
 const external = {
-  client,
+  client: savereturnClient,
   handleValidationError,
   resetUser,
   authenticate,
@@ -80,14 +86,14 @@ const external = {
 external.getEmailMessage = getMessage
 external.getSMSMessage = getMessage
 
-const sendEmail = async (_id, userData, dataArgs = {}, personalisation = {}) => {
+async function sendEmail (_id, userData, dataArgs = {}, personalisation = {}) {
   const message = external.getEmailMessage(_id, userData, dataArgs, personalisation)
-  return emailClient.sendMessage(message, {}, userData.logger)
+  return emailClient.send(message, {}, userData.logger)
 }
 
-const sendSMS = async (_id, userData, dataArgs = {}, personalisation = {}) => {
+async function sendSMS (_id, userData, dataArgs = {}, personalisation = {}) {
   const message = external.getSMSMessage(_id, userData, dataArgs, personalisation)
-  return smsClient.sendMessage(message, {}, userData.logger)
+  return smsClient.send(message, {}, userData.logger)
 }
 
 external.sendEmail = sendEmail
