+++ lib/module/savereturn/controller/savereturn.unit.spec.js
@@ -5,13 +5,13 @@ const { stub, useFakeTimers } = require('sinon')
 const proxyquire = require('proxyquire')
 
 const emailClient = {
-  sendMessage: () => {}
+  send: () => {}
 }
-const emailClientStub = stub(emailClient, 'sendMessage')
+const emailClientStub = stub(emailClient, 'send')
 const smsClient = {
-  sendMessage: () => {}
+  send: () => {}
 }
-const smsClientStub = stub(smsClient, 'sendMessage')
+const smsClientStub = stub(smsClient, 'send')
 
 const serviceData = require('~/fb-runner-node/service-data/service-data')
 const getInstanceStub = stub(serviceData, 'getInstance')
@@ -82,7 +82,7 @@ const userDetailsWithoutMobile = {
 }
 
 test('When requiring the savereturn module', t => {
-  const expectedClient = require('./client/save-return')
+  const expectedClient = require('~/fb-runner-node/module/savereturn/client/savereturn/savereturn')
   t.equal(client, expectedClient, 'it should export the save and return client')
   t.end()
 })
@@ -273,7 +273,7 @@ test('When sending emails', async t => {
   ], 'it should call getMessage with the expected args')
 
   t.ok(emailClientStub.calledOnce, 'it should send the email message')
-  t.deepEqual(emailClientStub.getCall(0).args, [{ _id: 'email.test' }, {}, userData.logger], 'it should call emailClient.sendMessage with the expected args')
+  t.deepEqual(emailClientStub.getCall(0).args, [{ _id: 'email.test' }, {}, userData.logger], 'it should call emailClient.send with the expected args')
 
   getEmailMessageStub.restore()
   t.end()
@@ -294,7 +294,7 @@ test('When sending sms messages', async t => {
   ], 'it should call getMessage with the expected args')
 
   t.ok(smsClientStub.calledOnce, 'it should send the sms message')
-  t.deepEqual(smsClientStub.getCall(0).args, [{ _id: 'sms.test' }, {}, userData.logger], 'it should call smsClient.sendMessage with the expected args')
+  t.deepEqual(smsClientStub.getCall(0).args, [{ _id: 'sms.test' }, {}, userData.logger], 'it should call smsClient.send with the expected args')
 
   getSMSMessageStub.restore()
   t.end()
