+++ lib/server/sources/configure/sources-configure-modules.js
@@ -11,11 +11,18 @@ const log = debug('runner:sources-configure-modules')
 debug.enable('runner:*')
 
 function configureModules (serviceConfigPath, runnerDir) {
+  log(serviceConfigPath, runnerDir)
+
   const serviceSources = []
   const modulesPath = path.join(runnerDir, 'lib/module')
   const modules = readdirSync(modulesPath).filter(dir => !dir.startsWith('.'))
 
+  log(modulesPath)
+
   const modulesConfigPath = path.join(serviceConfigPath, 'config.modules.json')
+
+  log(modulesConfigPath)
+
   let modulesConfig
   let modulesConfigRewrite
   try {
@@ -33,6 +40,9 @@ function configureModules (serviceConfigPath, runnerDir) {
   modules.forEach(module => {
     const moduleId = `config.module.${module}`
     const hasModuleConfig = modulesConfig.modules.filter(moduleConfig => moduleConfig._id === moduleId).length
+
+    log(moduleId, hasModuleConfig)
+
     if (!hasModuleConfig) {
       const modulePackagePath = path.join(modulesPath, module, 'package.json')
       const modulePackage = loadJson(modulePackagePath)
@@ -63,8 +73,12 @@ function configureModules (serviceConfigPath, runnerDir) {
     log(`Generated modules configuration file at "${modulesConfigPath}"`)
   }
 
+  log(modules)
+
   // Load any enabled modules
   modules.forEach(module => {
+    log(module)
+
     const moduleConfig = modulesConfig.modules.find((moduleConfig) => module === moduleConfig.module)
     if (moduleConfig.enabled !== 'on') {
       return
@@ -78,6 +92,8 @@ function configureModules (serviceConfigPath, runnerDir) {
       module
     })
   })
+
+  log(serviceSources)
 
   return serviceSources
 }
